refactor: main process
This completes the implementation of main process features. It features
basic window handling, data initialization and adjustments of decoders
to match the database and it's schemes.
A very rough renderer process script has been added to avoid depedencies
with errors to be used as well as to be able to test ipc.